First round at removing static httprequest layer, moving to HttpManager
